requirements.txt
pillow
torch
numpy
requests
huggingface_hub
segment_anything
diffusers
transformers
accelerate
scipy
safetensors
